Fix build error

+++ app/src/main/java/dev/app/tof/ToFFrame.kt
@@ -1,9 +1,9 @@
 package dev.app.tof
 
 data class ToFFrame(
-    val depth: ShortArray,   // 長度 = width*height
-    val amp: ShortArray,     // 同長度
+    val depth: IntArray,
+    val amp: IntArray,
     val width: Int,
     val height: Int,
-    val timestampNanos: Long = System.nanoTime()
+    val timestampNanos: Long = System.nanoTime()  // ← 給預設
 )
